changed paths \ to /, changed math type to double, removed rectangle.position and rectangle.size

+++ include/UtH/Math/Matrix4.hpp
@@ -2,8 +2,8 @@
 #ifndef MATRIX4_H_UMATH
 #define MATRIX4_H_UMATH
 
-#include <UtH\Math\Vector4.hpp>
-#include <UtH\Math\Matrix3.hpp>
+#include <UtH/Math/Vector4.hpp>
+#include <UtH/Math/Matrix3.hpp>
 
 namespace umath
 {
@@ -14,7 +14,7 @@ namespace umath
 		Creates Identity Matrix
 		*/
 		matrix4();
-		matrix4(const float* FirstElement);
+		matrix4(const double* FirstElement);
 		matrix4(const matrix3 mat3);
 		matrix4(
 			const vector4& Row1,
@@ -22,14 +22,14 @@ namespace umath
 			const vector4& Row3,
 			const vector4& Row4);
 		matrix4(
-			const float& r1s1,const float& r1s2,const float& r1s3,const float& r1s4,
-			const float& r2s1,const float& r2s2,const float& r2s3,const float& r2s4,
-			const float& r3s1,const float& r3s2,const float& r3s3,const float& r3s4,
-			const float& r4s1,const float& r4s2,const float& r4s3,const float& r4s4);
+			const double& r1s1,const double& r1s2,const double& r1s3,const double& r1s4,
+			const double& r2s1,const double& r2s2,const double& r2s3,const double& r2s4,
+			const double& r3s1,const double& r3s2,const double& r3s3,const double& r3s4,
+			const double& r4s1,const double& r4s2,const double& r4s3,const double& r4s4);
 		static const matrix4 Identity;
 		~matrix4();
 
-		float* FirstElement() const;
+		double* FirstElement() const;
 
 		const vector4& operator [](const unsigned int& index) const;
 		vector4& operator [](const unsigned int& index);
